+++ odoo/local-src/specific_base/models/res_company.py
@@ -52,7 +52,6 @@ class ResCompany(models.Model):
     active_campaigns = fields.Boolean(default=False,
                                       help="SEM active")
     form = fields.Char(string="Legal form", help="Juridical form")
-    company_name = fields.Char(help="Juridical name of the company")
     franchised_mobile = fields.Char()
     franchised_email = fields.Char()
     longitude = fields.Float()
@@ -92,6 +91,16 @@ class ResCompany(models.Model):
     royalties_freq = fields.Selection(
         selection='_get_royalties_freq_selection',
         default='1')
+
+    company_name = fields.Char('Center name')
+    center_street = fields.Char('Street')
+    center_street2 = fields.Char('Street2')
+    center_zip = fields.Char('ZIP')
+    center_city = fields.Char('City')
+    center_state_id = fields.Many2one('res.country.state', string='State')
+    center_country_id = fields.Many2one('res.country', string='Country')
+    center_email = fields.Char('Email')
+    center_phone = fields.Char('Phone')
 
     # product management
     can_create_product = fields.Boolean(default=False)