+++ odoo/songs/updates/update_10_1_1/clean_translation.sql
@@ -1,0 +1,16 @@
+-- Clean bad translation
+-- which will recreate correctly when the module will be updated
+
+DELETE FROM
+    ir_translation
+WHERE
+    module = 'sale_discount_program'
+AND lang = 'fr_FR'
+AND src = 'Quantity is the sum of matching products quantities';
+
+DELETE FROM
+    ir_translation
+WHERE
+    module = 'specific_base'
+AND lang = 'fr_FR'
+AND src = 'Company name';